update integration mode
