+++ includes/API.php
@@ -44,6 +44,63 @@ class API {
 		];
 
 		add_action( 'rest_api_init', [ $this, 'register_api' ] );
+		add_action( 'woocommerce_api_verify-bkash-payment', [ $this, 'verify_bkash_payment' ] );
+	}
+
+	public function verify_bkash_payment() {
+
+		$data = wp_unslash( $_GET );
+
+		$order = wc_get_order( $data['order_id'] );
+
+		// if it is not a valid order.
+		if ( ! $order instanceof \WC_Order ) {
+			wp_safe_redirect( site_url() );
+			exit;
+		}
+
+		// if the nonce is not valid.
+		if ( ! wp_verify_nonce( $data['nonce'], 'verify-bkash-payment' ) ) {
+			wp_safe_redirect( $order->get_checkout_order_received_url() );
+			exit;
+		}
+
+		// if the status is not success.
+		if ( 'success' !== $data['status'] ) {
+			wp_safe_redirect( $order->get_checkout_payment_url() );
+			exit;
+		}
+
+		$processor       = dc_bkash()->gateway->processor();
+		$execute_payment = $processor->execute_payment( $data['paymentID'] );
+
+		if ( is_wp_error( $execute_payment ) ) {
+			wp_safe_redirect( $order->get_checkout_payment_url() );
+			exit;
+		}
+
+		if ( $execute_payment ) {
+			do_action( 'dc_bkash_execute_payment_success', $order, $execute_payment );
+			wp_safe_redirect( $order->get_checkout_order_received_url() );
+			exit;
+		}
+
+
+//		array:12 [▼
+//		  "statusCode" => "0000"
+//		  "statusMessage" => "Successful"
+//		  "paymentID" => "TR00111I1679053331400"
+//		  "payerReference" => "dc_bkash64145212ee1ad"
+//		  "customerMsisdn" => "01717794286"
+//		  "trxID" => "ACH9JRYV7L"
+//		  "amount" => "2"
+//		  "transactionStatus" => "Completed"
+//		  "paymentExecuteTime" => "2023-03-17T17:42:47:127 GMT+0600"
+//		  "currency" => "BDT"
+//		  "intent" => "sale"
+//		  "merchantInvoiceNumber" => "69"
+//		]
+		exit;
 	}
 
 	/**
