+++ includes/Admin/Menu.php
@@ -54,7 +54,7 @@ class Menu {
 
 			$submenu[ $parent_slug ][] = [ __( 'Settings', 'dc-bkash' ), $capability, $this->get_submenu_url( 'settings' ) ]; // phpcs:ignore
 
-			$submenu[ $parent_slug ][] = [ __( 'Generate Doc', 'dc-bkash' ), $capability, $this->get_submenu_url( 'generate-doc' ) ]; // phpcs:ignore
+			//$submenu[ $parent_slug ][] = [ __( 'Generate Doc', 'dc-bkash' ), $capability, $this->get_submenu_url( 'generate-doc' ) ]; // phpcs:ignore
 		}
 
 		add_action( 'load-' . $hook, [ $this, 'init_hooks' ] );
